requirements.txt
flask==3.0.0
gunicorn==21.2.0
firebase-admin==6.2.0
openai>=1.50.0
openai-agents
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
pytest==7.4.0
faker==19.3.0